pyproject.toml
[tool.poetry]
name = "BounceRL"
version = "0.0.50"
description = "Bounce RL"
authors = ["William Henning <wchenningwchenning@gmail.com>"]
readme = "README.md"
packages = [
    { include = "bounce_rl", from = "."},
]

[tool.poetry.dependencies]
python = ">=3.10"
numpy = "^2.2.0"
cffi = "^1.16.0"
simplejpeg = "^1.7.0"
psutil = "^5.9.8"
python-xlib = "^0.33"
libtimecontrol = "^0.2.1"
sphinx = "^7.3.0"
sphinx-autodoc-typehints = "^2.3.0"
sphinx-autobuild = "^2024.10.3"
furo = "^2025.7.19"
bounce-desktop = "^0.2.2"
factorio-state-exporter = {git = "https://github.com/Whenning42/factorio-state-exporter.git"}
gymnasium = "^1.2.2"
pytest = "^9.0.2"
pillow = "^12.0.0"

[build-system]
requires = [
    "poetry-core>=1.2.1",
    "setuptools",
    "numpy",
    "cython",
]
build-backend = "poetry.core.masonry.api"
